Apply CUD repo to update Acc/Bcc

+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/data/node/CcAsccpNodeDetail.java
@@ -13,8 +13,8 @@ public class CcAsccpNodeDetail implements CcNodeDetail {
 
     @Data
     public static class Ascc {
-        private long manifestId;
-        private long asccId;
+        private BigInteger manifestId;
+        private BigInteger asccId;
         private String guid;
         private String den;
         private int cardinalityMin;
